requirements.txt
beautifulsoup4==4.12.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
lxml==5.3.0
selectolax==0.3.21
click==8.1.7
pandas==2.2.3
urllib3==1.26.18
requests==2.31.0
plotly==5.22.0
python-igraph==0.11.5